feat: #3 채contents-request

+++ src/main/java/com/yourmode/yourmodebackend/domain/request/dto/user/response/ContentRequestSummaryDto.java
@@ -1,5 +1,6 @@
 package com.yourmode.yourmodebackend.domain.request.dto.user.response;
 
+import com.yourmode.yourmodebackend.domain.request.dto.user.request.ContentRequestStatusHistoryDto;
 import lombok.*;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -11,7 +12,7 @@ import java.util.List;
 @Builder
 public class ContentRequestSummaryDto {
 
-    private Integer id;
+    private Long id;
     private String bodyFeature;
     private String situation;
     private Integer budget;
